// src/components/Card/types.ts
interface User {
    avatar_url?: string;
    name?: string;
    public_repos?: number;
    followers?: number;
    following?: number;
    location?: string;
    blog?: string;
    company?: string;
    twitter_username?: string;
    bio?: string;
    created_at?: string;
    login?: string;
}

export type { User }